website: move mobile nav to own file
